+++ polygon_coverage_geometry/CMakeLists.txt
@@ -29,10 +29,30 @@ target_link_libraries(${PROJECT_NAME} ${CGAL_LIBRARIES} ${CGAL_3RD_PARTY_LIBRARI
 #########
 # TESTS #
 #########
+catkin_add_gtest(test_bcd
+  test/bcd-test.cpp
+)
+target_link_libraries(test_bcd ${PROJECT_NAME})
+
+catkin_add_gtest(test_cgal_comm
+  test/cgal_comm-test.cpp
+)
+target_link_libraries(test_cgal_comm ${PROJECT_NAME})
+
 catkin_add_gtest(test_offset
   test/offset-test.cpp
 )
 target_link_libraries(test_offset ${PROJECT_NAME})
+
+catkin_add_gtest(test_sweep
+  test/sweep-test.cpp
+)
+target_link_libraries(test_sweep ${PROJECT_NAME})
+
+catkin_add_gtest(test_visibility_polygon
+  test/visibility_polygon-test.cpp
+)
+target_link_libraries(test_visibility_polygon ${PROJECT_NAME})
 
 ##########
 # EXPORT #
